# documentation/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Node js test api",
        "description": "Ventaja Government Pay Test apis node js",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "contact": {
            "name": "appscrip",
            "email": "jagannath@appscrip.co"
        }
    },
    "components": {
        "securitySchemes": [
            {
                "bearerAuth": {
                    "type": "http",
                    "description": "access token",
                    "scheme": "bearer",
                    "bearerFormat": "JWT",
                    "in": "header"
                }
            }
        ]
    },
    "security": {},
    "servers": [
        "http://localhost:8000/"
    ],
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/bill-payment/outlet-balance": {
            "post": {
                "tags": [
                    "Bill Payment"
                ],
                "summary": "Outlet balance of bill payment",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "required": [
                                "*"
                            ]
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "properties": {
                                "status": "200",
                                "message": "success",
                                "data": {
                                    "statusCode": "",
                                    "data": ""
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/400"
                        }
                    },
                    "500": {
                        "schema": {
                            "$ref": "#/definitions/500"
                        }
                    }
                }
            }
        },
        "/bill-payment/validate-topup": {
            "post": {
                "tags": [
                    "Bill Payment"
                ],
                "summary": "bill payment validate topup",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "required": [
                                "*"
                            ],
                            "properties": {
                                "payremitNo": {
                                    "type": "string",
                                    "example": "1017108390499970"
                                },
                                "amount": {
                                    "type": "number",
                                    "example": 1000.0
                                },
                                "currencyCode": {
                                    "type": "string",
                                    "example": "PHP"
                                },
                                "agentRefNo": {
                                    "type": "string",
                                    "example": "A1234"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "properties": {
                                "status": "200",
                                "message": "success",
                                "data": {
                                    "statusCode": "",
                                    "data": ""
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/400"
                        }
                    },
                    "500": {
                        "schema": {
                            "$ref": "#/definitions/500"
                        }
                    }
                }
            }
        },
        "/bill-payment/process-topup": {
            "post": {
                "tags": [
                    "Bill Payment"
                ],
                "summary": "bill payment process topup",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "required": [
                                "*"
                            ],
                            "properties": {
                                "payremitNo": {
                                    "type": "string",
                                    "example": "1017108390499970"
                                },
                                "amount": {
                                    "type": "number",
                                    "example": 1000.0
                                },
                                "currencyCode": {
                                    "type": "string",
                                    "example": "PHP"
                                },
                                "agentRefNo": {
                                    "type": "string",
                                    "example": "A1234"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "properties": {
                                "status": "200",
                                "message": "success",
                                "data": {
                                    "statusCode": "",
                                    "data": ""
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/400"
                        }
                    },
                    "500": {
                        "schema": {
                            "$ref": "#/definitions/500"
                        }
                    }
                }
            }
        },
        "/bill-payment/topup-transaction-inquiry": {
            "post": {
                "tags": [
                    "Bill Payment"
                ],
                "summary": "bill payment topup transaction inquiry",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "required": [
                                "*"
                            ],
                            "properties": {
                                "transactionNumber": {
                                    "type": "string",
                                    "example": "PRT-13202010-000001"
                                },
                                "referenceNumber": {
                                    "type": "string",
                                    "example": "TU-202010000001"
                                },
                                "agentRefNo": {
                                    "type": "string",
                                    "example": "A1234"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "properties": {
                                "status": "200",
                                "message": "success",
                                "data": {
                                    "statusCode": "",
                                    "data": ""
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/400"
                        }
                    },
                    "500": {
                        "schema": {
                            "$ref": "#/definitions/500"
                        }
                    }
                }
            }
        },
        "/bill-payment/process-reservation-detail": {
            "post": {
                "tags": [
                    "Bill Payment"
                ],
                "summary": "bill payment process reservation detail",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "required": [
                                "*"
                            ],
                            "properties": {
                                "reservationNumber": {
                                    "type": "string",
                                    "example": "RN-202007000011"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "properties": {
                                "status": "200",
                                "message": "success",
                                "data": {
                                    "statusCode": "",
                                    "data": ""
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/400"
                        }
                    },
                    "500": {
                        "schema": {
                            "$ref": "#/definitions/500"
                        }
                    }
                }
            }
        },
        "/bill-payment/validate-payment-reservation-detail": {
            "post": {
                "tags": [
                    "Bill Payment"
                ],
                "summary": "bill payment validate payment reservation detail",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "required": [
                                "*"
                            ],
                            "properties": {
                                "reservationNumber": {
                                    "type": "string",
                                    "example": "RN-202007000011"
                                },
                                "amount": {
                                    "type": "number",
                                    "example": 150
                                },
                                "currencyCode": {
                                    "type": "string",
                                    "example": "USD"
                                },
                                "settlementCurrencyCode": {
                                    "type": "string",
                                    "example": "PHP"
                                },
                                "settlementAmount": {
                                    "type": "number",
                                    "example": 8250
                                },
                                "agentRefNo": {
                                    "type": "string",
                                    "example": "A1234"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "properties": {
                                "status": "200",
                                "message": "success",
                                "data": {
                                    "statusCode": "",
                                    "data": ""
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/400"
                        }
                    },
                    "500": {
                        "schema": {
                            "$ref": "#/definitions/500"
                        }
                    }
                }
            }
        },
        "/bill-payment/process-payment-reservation-detail": {
            "post": {
                "tags": [
                    "Bill Payment"
                ],
                "summary": "bill payment process payment reservation detail",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "required": [
                                "*"
                            ],
                            "properties": {
                                "reservationNumber": {
                                    "type": "string",
                                    "example": "RN-202007000011"
                                },
                                "amount": {
                                    "type": "number",
                                    "example": 150
                                },
                                "currencyCode": {
                                    "type": "string",
                                    "example": "USD"
                                },
                                "settlementCurrencyCode": {
                                    "type": "string",
                                    "example": "PHP"
                                },
                                "settlementAmount": {
                                    "type": "number",
                                    "example": 8250
                                },
                                "agentRefNo": {
                                    "type": "string",
                                    "example": "A1234"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "properties": {
                                "status": "200",
                                "message": "success",
                                "data": {
                                    "statusCode": "",
                                    "data": ""
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/400"
                        }
                    },
                    "500": {
                        "schema": {
                            "$ref": "#/definitions/500"
                        }
                    }
                }
            }
        },
        "/bill-payment/reservation-payment-inquiry": {
            "post": {
                "tags": [
                    "Bill Payment"
                ],
                "summary": "bill payment reservation payment inquiry",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "required": [
                                "*"
                            ],
                            "properties": {
                                "reservationNumber": {
                                    "type": "string",
                                    "example": "RN-202007000011"
                                },
                                "amount": {
                                    "type": "number",
                                    "example": 150
                                },
                                "currencyCode": {
                                    "type": "string",
                                    "example": "USD"
                                },
                                "settlementCurrencyCode": {
                                    "type": "string",
                                    "example": "PHP"
                                },
                                "settlementAmount": {
                                    "type": "number",
                                    "example": 8250
                                },
                                "agentRefNo": {
                                    "type": "string",
                                    "example": "A1234"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "properties": {
                                "status": "200",
                                "message": "success",
                                "data": {
                                    "statusCode": "",
                                    "data": ""
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/400"
                        }
                    },
                    "500": {
                        "schema": {
                            "$ref": "#/definitions/500"
                        }
                    }
                }
            }
        },
        "/bill-payment/pull-transaction-detail": {
            "post": {
                "tags": [
                    "Bill Payment"
                ],
                "summary": "bill payment pull transaction detail",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "required": [
                                "*"
                            ],
                            "properties": {
                                "code": {
                                    "type": "number",
                                    "example": 106
                                },
                                "transactionNumber": {
                                    "type": "string",
                                    "example": "PRP-13202007-000006"
                                },
                                "referenceNumber": {
                                    "type": "string",
                                    "example": "RP-202007000001"
                                },
                                "agentRefNo": {
                                    "type": "string",
                                    "example": "A1234"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "properties": {
                                "status": "200",
                                "message": "success",
                                "data": {
                                    "statusCode": "",
                                    "data": ""
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/400"
                        }
                    },
                    "500": {
                        "schema": {
                            "$ref": "#/definitions/500"
                        }
                    }
                }
            }
        },
        "/bill-payment/validate-billing": {
            "post": {
                "tags": [
                    "Bill Payment"
                ],
                "summary": "bill payment validate billing",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "required": [
                                "*"
                            ],
                            "properties": {
                                "billerCode": {
                                    "type": "string",
                                    "example": "C15"
                                },
                                "accountNumber": {
                                    "type": "string",
                                    "example": "816218000"
                                },
                                "identifier": {
                                    "type": "string",
                                    "example": "Juan Dela Cruz"
                                },
                                "amount": {
                                    "type": "number",
                                    "example": 120
                                },
                                "agentRefNo": {
                                    "type": "string",
                                    "example": "A1234"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "properties": {
                                "status": "200",
                                "message": "success",
                                "data": {
                                    "statusCode": "",
                                    "data": ""
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/400"
                        }
                    },
                    "500": {
                        "schema": {
                            "$ref": "#/definitions/500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "400": {
            "type": "object",
            "properties": {
                "responseCode": {
                    "type": "number",
                    "example": 131
                },
                "message": {
                    "type": "string",
                    "example": "Duplicate transaction"
                },
                "description": {
                    "type": "string",
                    "example": "Applicable date has an overlapping payment"
                },
                "status": {
                    "type": "number",
                    "example": 400
                }
            }
        },
        "500": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Internal server error!"
                },
                "status": {
                    "type": "integer",
                    "example": 500
                }
            }
        }
    }
}